+++ function.php
@@ -1,7 +1,7 @@
 <?php
 
 // Soft version
-define('VERSION', '3.2');
+define('VERSION', '3.4');